[refactor] 데이터 Fetch의 성공 여부를 파악하는 Result, 그리고 응답 상태를 나타내는 Status
- 근데 UI Layer에 포함되어야 하는지 잘 모르겠음.
