+++ app/src/main/java/com/bodan/maplecalendar/presentation/utils/Status.kt
@@ -1,0 +1,5 @@
+package com.bodan.maplecalendar.presentation.utils
+
+enum class Status {
+    SUCCESS, ERROR, FAIL, LOADING
+}
